[refactor/#43] ReservationInfo 접근성 코드 정리
